[nodejs-review] review-3 fixed. move test code to src dir from test dir
