+++ projects/koajs-proj-conf/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "lint": "npx eslint ./app.js ./server.js ./src",
     "start": "npx nodemon ./server.js",
-    "test": "NODE_ENV=test jest"
+    "test": "NODE_ENV=test jest --env=node"
   },
   "author": "",
   "license": "MIT",
